Add token config parameter
This is a rough draft of allowing the provider to accept a token instead
of using username and password. The intention is that the user would
retrieve an auth token, possibly extend the expiration, outside of
Terraform. This allows for multiple providers to use the same token to
avoid exhausting the number of tokens from the BigIP.