Add full_path to datasource_bigip_ltm_node
Add test to cover the data source. The original test was not testing the
data source directly, so I updated the test to compare the resource and
data source.

Add `full_path` to the `bigip_ltm_node data source`. The reason for this
addition is that `name` and `id` do not include the Partition as most
other resources do. I added `full_path` because I was worried about
breaking people who are already using the bigip_ltm_node data source.
The alternative would have been to fix this `id` and `name` properties
at the risk of breaking consumers of the provider.

add full_path to data source docmentation